ASM-3830 Rack switches not getting configured when PowerConnect N in inventory
PowerConnect switch discovery code is using mac-address table information for remote-host fact. It is observed that mac-address table is listing the mac address of equipment connect to other switches via upstream port-channel.

These switches are not used for end to end deployment with ASM, For now removing the remote host information, till a better solution is identified.